treez_assignment/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==17.0.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
